+++ src/types/internal.ts
@@ -13,6 +13,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+export type ClassOrInstance<T extends any = Object> = T | Constructor<T>;
+
 export type ClassPropKey = string | symbol;
 
 export type Collection<T extends any = any> =