+++ src/utils/internal.ts
@@ -13,13 +13,29 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import type { Collection } from "../types/internal";
+import type { ClassOrInstance, Collection, Constructor, Nullable } from "../types/internal";
+
+export function getClassName(classOrInstance: ClassOrInstance<any>): Nullable<string> {
+    if (isConstructor(classOrInstance)) {
+        return classOrInstance.name;
+    }
+
+    if (typeof classOrInstance?.constructor?.name === "string") {
+        return classOrInstance?.constructor?.name;
+    }
+
+    return null;
+}
 
 export function isCollection<T extends any = any>(val: any): val is Collection<T> {
     return !!val && (
         "push" in val ||
         "add" in val
     );
+}
+
+export function isConstructor<T extends any = any>(classOrInstance: ClassOrInstance<T>): classOrInstance is Constructor<T> {
+    return typeof classOrInstance === "function";
 }
 
 export function isNil(val: unknown): val is (null | undefined) {