+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (@egomobile/documentation)
 
+## 0.3.0
+
+- implement serializable types and `serializeDependencies()` utility function
+- add `classOrInstance` prop to `IMethodDependencyItem`, `IParameterDependencyItem` and `IPropertyDependencyItem`
+- add optional `key` to `IDependencyInformation`
+
 ## 0.2.0
 
 - add `references` props to `IDependencyInformation`, `IDependencyInformationEntity` and `IDependencyInformationEntityAttribute`